+++ src/lib.rs
@@ -33,8 +33,8 @@ use metrics::MetricInfoWrapper;
 
 use std::collections::HashMap;
 use std::convert::Infallible;
+use std::env;
 use std::time::Duration;
-use std::{env, str::Utf8Error};
 
 use base64::Engine;
 use http_body_util::combinators::BoxBody;
@@ -42,11 +42,11 @@ use http_body_util::{Empty, Full};
 use hyper::body::Bytes;
 use hyper::{body::Buf, Method};
 use hyper_util::client::legacy::{connect::HttpConnector, Builder, Client};
-use quick_error::quick_error;
 use serde::{Deserialize, Serialize};
 use slog_scope::{error, info};
 use tokio::time::timeout;
 
+use errors::*;
 #[cfg(feature = "metrics")]
 use http::StatusCode;
 
@@ -64,67 +64,14 @@ pub use metrics::MetricInfo;
 pub use metrics::{Function, HttpMethod};
 pub use types::*;
 
+/// Consul errors
+pub mod errors;
 #[cfg(feature = "trace")]
 mod hyper_wrapper;
 /// Types exposed for metrics on the consuming application without taking a dependency on a metrics library or a specific implementation.
 mod metrics;
 /// The strongly typed data structures representing canonical consul objects.
 pub mod types;
-
-quick_error! {
-    /// The error type returned from all calls into this this crate.
-    #[derive(Debug)]
-    pub enum ConsulError {
-        /// The request was invalid and could not be serialized to valid json.
-        InvalidRequest(err: serde_json::error::Error) {}
-        /// The request was invalid and could not be converted into a proper http request.
-        RequestError(err: http::Error) {}
-        /// The consul server response could not be converted into a proper http response.
-        ResponseError(err: hyper_util::client::legacy::Error) {}
-        /// The consul server response was invalid.
-        InvalidResponse(err: hyper::Error) {}
-        /// The consul server response could not be deserialized from json.
-        ResponseDeserializationFailed(err: serde_json::error::Error) {}
-        /// The consul server response could not be deserialized from bytes.
-        ResponseStringDeserializationFailed(err: std::str::Utf8Error) {}
-        /// The consul server response was something other than 200. The status code and the body of the response are included.
-        UnexpectedResponseCode(status_code: hyper::http::StatusCode, body: Option<String>) {}
-        /// The consul server refused a lock acquisition (usually because some other session has a lock).
-        LockAcquisitionFailure(err: u64) {}
-        /// Consul returned invalid UTF8.
-        InvalidUtf8(err: Utf8Error) {
-            from()
-        }
-        /// Consul returned invalid base64.
-        InvalidBase64(err: base64::DecodeError) {
-            from()
-        }
-        /// IO error from sync api.
-        SyncIoError(err: std::io::Error) {
-            from()
-        }
-        /// Response parse error from sync api.
-        SyncInvalidResponseError(err: std::str::ParseBoolError) {
-            from()
-        }
-        /// Unexpected response code from sync api.
-        SyncUnexpectedResponseCode(status_code: u16, body: String) {}
-        /// Consul request exceeded specified timeout.
-        TimeoutExceeded(timeout: std::time::Duration) {
-            display("Consul request exceeded timeout of {:?}", timeout)
-        }
-        /// Unable to resolve the service's instances in Consul.
-        ServiceInstanceResolutionFailed(service_name: String) {
-            display("Unable to resolve service '{}' to a concrete list of addresses and ports for its instances via consul.", service_name)
-        }
-        /// An error from ureq occured.
-        UReqError(err: ureq::Error) {
-            display("UReq error: {}", err)
-        }
-    }
-}
-
-pub(crate) type Result<T> = std::result::Result<T, ConsulError>;
 
 /// The config necessary to create a new consul client.
 #[derive(Clone, Debug, Serialize, Deserialize)]
